Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scripthash-malformed.html

<script>
    if (window.testRunner)
        testRunner.dumpAsText();
</script>
<meta http-equiv="Content-Security-Policy" content="script-src 'sha1-vMUNaexq6wlYKxWWPlJUOvhn67U=' 'sha256-j+ToFnhur5mTSCwoyy4Fh; 'sha512-0tQJUSPHRGUYQIa2ByQWZSlUa5sI3e2tOEXF7sg5mjmsHk/EHDTTORIDyMPhDKgCmb7TIrwhLrq2kEvfjHQ=='">
